* Added simple XML binder for handling XML requests and responses and translating them into C structures.  Code is generated directly from the XSD file(s).  * Added Atmos 2.1 anonymous access token support.
git-svn-id: https://atmos-c.googlecode.com/svn/branches/dev-atmos-c-2.1@91 cb61ccbc-2dfe-265f-a3f4-a677fceac7d0

+++ lib/atmos_read.c
@@ -1154,9 +1154,9 @@ void AtmosFilter_get_listable_tags(RestFilter *self, RestClient *rest,
         const char *comma = strchr(tag, ',');
         if(!comma) {
             // End of string reached.  Take what's left.
-            c = strlen(tag);
+            c = (int)strlen(tag);
         } else {
-            c = comma - tag;
+            c = (int)(comma - tag);
         }
         if(res->tag_count + 1 > tags_alloc) {
             char **newalloc;
